add counter decrement method
this commit does a couple of things. First of all it adds methods to decrement
counters. Previously it was only possible to increment as passing in a
negative number would overflow uint64_t.

In addition to that, over and underflow protection was added to the increment
and decrement methods to make sure we stay in range when doing calculations on
behalf of the caller. The gauge method doesn't have this as it is the
responsibility of the caller to pass in the right value.

Lastly this commit changes the return values of all recording methods to `int`
as it makes more sense than `size_t` since we don't handle any size related
things.